+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.0.4 [unreleased]
 
-- Added `word: false` option
 - Added `match` option
+- Added `word: false` option
 - Added highlighted fields to `load: false`
 
 ## 1.0.3